fixed textview not being clickable and made album titles scrollable
Making the text scrollable took away the ripple animations and it’s
ability to field the onclicklistener, so I had to recreate and
implement ripple animations and on touch from scratch. It works,
although not as smoothly as I’d like. Test it out and let me know your
thoughts. I’m not sure there’s an easier way to do this, but someone
can try and find one.