Added ItemDecoration to Songs,Ablums,Artist and Playlists
Since the 4 fragments are based on the
AbsLibaryPageRecyclerViewFragment, I added the decoration code there,
since the recycler view setup is also there. Had to da a bit of
workaround for the PlaylistFragment and the SongsFragment, hence the
shouldDrawLine method. The check for adapter is there, because the
default viewtype is 0 and the item would't draw the line on the other
lists. There was also a bit of work regarding removing/readding the
divider because of the layout item switching.