Legacy notch support
Phonograph draws the status bar artificially (fixed-sized, darker colored background).
The default status bar size is set (correctly!) to 25dp on pre-M devices and 24dp on M+ devices.
Since some devices nowadays have notches the status bar size on those phones differs from the default.

On OnePlus 6 the status bar matches the size of the notch, causing the old implementation to draw a smaller "status bar" leading to weird UI. On OP6 the status bar size is 30+dp. Since the size is a system defined property it should work on all devices regardless of the presence of a notch.

This fix changes the height of the status bar from fixed size to the system defined property.
This fix is not perfect as it is accessing a private property but ATM there is no official API that can be used instead. It will work as long as the property name doesn't change on OS level. Proper test would be to run the app on different APIs.

Tested: OnePlus 6, Oreo 8.1